reorganizing + using PRODUCTION_TENANTS instead of a useCustomDomain attribute on tenants collection + updating env vars
