+++ .env.example
@@ -12,5 +12,15 @@ PAYLOAD_SECRET=YOUR_SECRET_HERE
 # Used to configure CORS, format links and more. No trailing slash
 NEXT_PUBLIC_ROOT_DOMAIN=localhost:3000
 
+# A comma separated list of tenant slugs. Tenants in this list will have their customDomain used for routing.
+# You likely want to leave this blank in your local env.
+PRODUCTION_TENANTS=
+
 # The API key for the Google Maps SDK we pass to the National Avalanche Center widgets.
 NEXT_PUBLIC_GOOGLE_MAPS_API_KEY=18514e92a54f6afc9a7fe83957be4fc3f1a89328
+
+# avalanche.org API base url. Defaults to https://api.avalanche.org. Not required.
+NAC_HOST=
+
+# AFP wordpress API base url. Defaults to https://forecasts.avalanche.org. Not required.
+AFP_HOST=
