+++ src/app/(frontend)/[center]/layout.tsx
@@ -9,8 +9,8 @@ import { mergeOpenGraph } from '@/utilities/mergeOpenGraph'
 import { AvalancheCenterProvider } from '@/providers/AvalancheCenterProvider'
 import { TenantProvider } from '@/providers/TenantProvider'
 import { getAvalancheCenterMetadata, getAvalancheCenterPlatforms } from '@/services/nac/nac'
-import { getHostnameFromTenant } from '@/utilities/getHostnameFromTenant'
 import { getURL } from '@/utilities/getURL'
+import { getHostnameFromTenant } from '@/utilities/tenancy/getHostnameFromTenant'
 import { cn } from '@/utilities/ui'
 import configPromise from '@payload-config'
 import { getPayload } from 'payload'
