+++ src/environment.d.ts
@@ -3,8 +3,8 @@ declare global {
     interface ProcessEnv {
       PAYLOAD_SECRET: string
       DATABASE_URI: string
-      NEXT_PUBLIC_SERVER_URL: string
-      VERCEL_PROJECT_PRODUCTION_URL: string
+      NEXT_PUBLIC_ROOT_DOMAIN: string
+      PRODUCTION_TENANTS: string
       NAC_HOST: string
       AFP_HOST: string
       NEXT_PUBLIC_GOOGLE_MAPS_API_KEY: string
