+++ src/utilities/generatePreviewPath.ts
@@ -1,5 +1,6 @@
 import { Tenant } from '@/payload-types'
 import { CollectionSlug, PayloadRequest } from 'payload'
+import { getURL } from './getURL'
 
 const collectionPrefixMap: Partial<Record<CollectionSlug, string>> = {
   posts: '/posts',
@@ -36,11 +37,7 @@ export const generatePreviewPath = ({ collection, slug, tenant, req }: Props) =>
     encodedParams.append(key, value)
   })
 
-  const isProduction =
-    process.env.NODE_ENV === 'production' || Boolean(process.env.VERCEL_PROJECT_PRODUCTION_URL)
-  const protocol = isProduction ? 'https:' : req.protocol
-
-  const url = `${protocol}//${currentHost}/${shouldIncludeTenantSlug ? `${tenant.slug}/` : ''}next/preview?${encodedParams.toString()}`
+  const url = `${getURL(currentHost)}/${shouldIncludeTenantSlug ? `${tenant.slug}/` : ''}next/preview?${encodedParams.toString()}`
 
   return url
 }
