+++ src/utilities/tenancy/getHostnameFromTenant.ts
@@ -1,10 +1,11 @@
 import { Tenant } from '@/payload-types'
-import { ROOT_DOMAIN } from './domain'
+import { ROOT_DOMAIN } from '../domain'
+import { getProductionTenantSlugs } from './getProductionTenants'
 
 export function getHostnameFromTenant(tenant: Tenant | null) {
   if (!tenant) return ROOT_DOMAIN
 
-  if (tenant.useCustomDomain) {
+  if (getProductionTenantSlugs().includes(tenant.slug)) {
     return tenant.customDomain
   }
