+++ src/utilities/tenancy/getProductionTenants.ts
@@ -1,0 +1,3 @@
+export function getProductionTenantSlugs() {
+  return (process.env.PRODUCTION_TENANTS || '').split(',').map((str) => str.trim())
+}
